refactor(engine): move CommandResponseWriter into api
